homework9/services/billing/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2